==> 1. Dictionary Exploration.ipynb <==
# Dictionary Exploration
This notebook explores the files in `../Dataset/cv-corpus-20.0-delta-2024-12-06/en`.

> **Conclusion:** The information in the dictionaries is limited and useless. There is no way to match the dictionaries with the audios.

import os
import pandas as pd
import numpy as np
import json

file_path = "../Dataset/cv-corpus-20.0-delta-2024-12-06/en/cv-corpus-20.0-2024-12-06.json"

with open(file_path, 'r') as f:
    data = json.load(f)

np.sum([i for i in data["locales"]["en"]["buckets"].values()])

data["locales"]["en"]

## Dictionaries
- validated_sentences.tsv
- unvalidated_sentences.tsv
- validated.tsv
- clip_durations.tsv
- other.tsv
- reported.tsv
- invalidated.tsv

validated_sentences = pd.read_csv(data_path + "validated_sentences.tsv", sep="\t")
validated_sentences.shape

validated_sentences.iloc[0]

# We will work only with validates sentences
validated_sentences_ids = validated_sentences["sentence_id"].values

unvalidated_sentences = pd.read_csv(data_path + "unvalidated_sentences.tsv", sep="\t")
unvalidated_sentences.shape

unvalidated_sentences.iloc[0]

# This file could be replaced with other.tsv
validated = pd.read_csv(data_path + "validated.tsv", sep="\t")
print(validated.shape) # Only 250 audios are in this dataset. The 250 audios present in validated_sentences

validated.iloc[0]

clip_durations = pd.read_csv(data_path + "clip_durations.tsv", sep="\t")
clip_durations.shape

clip_durations.iloc[0] # It's not clear how we can match the information between files!

other = pd.read_csv(data_path + "other.tsv", sep="\t")
other.shape

other.iloc[0] 

# We have age information for 75% of the data and gender information for 43%
# sentence_domain, variant and segment are useless.
other.isna().mean()

# All the files in other are in validated_sentences
print(len(set(other["sentence_id"].values) - set(validated_sentences_ids)))
# validated_sentences has 1'648.601 files that are not in other
len(set(validated_sentences_ids) - set(other["sentence_id"].values))

reported = pd.read_csv(data_path + "reported.tsv", sep="\t")
reported.shape

reported.iloc[0]

invalidated = pd.read_csv(data_path + "invalidated.tsv", sep="\t")
invalidated.shape

invalidated.iloc[0]

# Cleaning validated sentences
corrupt_ids = pd.concat([invalidated["sentence_id"], reported["sentence_id"]]).reset_index(drop = True)
corrupt_ids = corrupt_ids[corrupt_ids.notna()].values

# New validated_sentences_ids
validated_sentences_ids = list(set(validated_sentences_ids) - set(corrupt_ids))
